data discovery/requirements.txt
boto3>=1.26.0
azure-storage-blob>=12.14.0
azure-identity>=1.12.0
google-cloud-storage>=2.7.0
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-datatransfer>=1.11.0
google-cloud-bigquery-reservation>=1.11.0
pymongo>=4.3.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
sqlalchemy>=1.4.0
pandas>=1.5.0
pyarrow>=10.0.0
openpyxl>=3.0.0
paramiko>=2.12.0
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=0.19.0
fastapi>=0.104.0
uvicorn>=0.24.0
smbprotocol>=1.10.0
pysmb>=1.2.8
watchdog>=3.0.0

# Additional Cloud Providers
oci>=2.100.0
ibm-cloud-sdk-core>=3.16.0
ibm-boto3>=1.0.0
ibmcloudant>=0.5.0
ibm-watson>=6.3.0
ibm-cloud-resource-manager>=1.0.0
alibabacloud_oss2>=2.17.0
alibabacloud_rds20140815>=3.0.0
alibabacloud_tea_openapi>=0.3.0

# Additional Database Connectors
cx_Oracle>=8.3.0
pyodbc>=4.0.0
cassandra-driver>=3.25.0
neo4j>=5.5.0
redis>=4.5.0
elasticsearch>=8.6.0
snowflake-connector-python>=3.0.0
influxdb>=5.3.1

# Data Warehouse & Analytics
databricks-sql-connector>=2.5.0
teradatasql>=17.20.0
clickhouse-driver>=0.2.5
presto-python-client>=0.8.4
trino>=0.324.0
apache-superset>=2.1.0

# Streaming & Message Queues
kafka-python>=2.0.2
confluent-kafka>=2.0.2
azure-eventhub>=5.11.0
azure-servicebus>=7.8.0
pika>=1.3.0
pulsar-client>=3.1.0
nats-py>=2.3.0
google-cloud-pubsub>=2.18.0

# SaaS Platforms - CRM
simple-salesforce>=1.12.0
hubspot-api-client>=7.0.0
pipedrive-python-lib>=1.1.0
zoho-crm-python-sdk>=3.0.0

# SaaS Platforms - Support & Ticketing
pysnow>=0.7.17
zenpy>=2.0.25
freshdesk-python>=1.0.0
intercom-python>=3.1.0

# SaaS Platforms - HR & Finance
workday>=0.1.0
bamboohr>=1.0.0
adp-python>=1.0.0
quickbooks-python>=1.0.0
xero-python>=3.0.0
sage-python>=1.0.0

# SaaS Platforms - Communication
jira>=3.4.0
slack-sdk>=3.20.0
discord-py>=2.3.0
microsoft-graph-python>=1.0.0

# SaaS Platforms - Marketing & Analytics
mailchimp-marketing>=3.0.80
marketo-python>=1.0.0
google-analytics-reporting>=0.7.0
adobe-analytics-python>=1.0.0
mixpanel-python>=4.10.0

# Data Lake & Modern Formats
deltalake>=0.10.0
pyiceberg>=0.5.0
pyhudi>=0.1.0
pyarrow>=10.0.0
minio>=7.1.0
hdfs3>=0.3.1
pydoop>=1.3.0

# Additional Time Series & Analytics
timescaledb-python>=1.0.0
questdb-python>=1.0.0
influxdb-client>=1.36.0
apache-druid-python>=1.0.0
apache-pinot-python>=1.0.0
clickhouse-connect>=0.6.0

# Additional Graph & Document DBs
arangodb-python-driver>=7.5.0
faunadb>=4.5.1
couchdb>=1.2.0
rethinkdb>=2.4.10

# Additional Search & Cache
opensearch-py>=2.3.0
apache-solr-python>=1.0.0
hazelcast-python-client>=5.2.0

# Regional Cloud Providers
tencentcloud-sdk-python>=3.0.0
baidubce>=0.9.0
huaweicloudsdkcore>=3.1.0
volcengine-python-sdk>=1.0.0

# Specialized Storage & File Systems
glusterfs-python>=1.0.0
cephfs>=1.0.0
swift-python>=1.0.0
